Beskrivelse av kvalitetsindikatorer

Oversikt over over og beskrivelse av alle kvalitetsindikatorerer. Fila oppdateres etter hvert som det legges til data fra flere kvalitetsindikatorer og registre. "Originalfila" finnes som Excel-fil under data-raw. Gjør endringer der og skriv så ned til rda.

IndBeskr

Format

Dataramme med ei rad per indikator. Som variabler er ulike beskrivende elementer for indikatoren

IndID

Indikatorens iD. Består av registernavn og et løpenummer. Eks nkr_nakke2, intensiv1

Register

kortnavn for registeret. Eks nakke, intensiv. Denne er strengt tatt overflødig hvis man er nøye med konsekvent navngiving av indikatorens id, se IndID

IndTittel

Tittel eller en form for kort beskrivelse av indikatoren

IndNavn

Indikatornavn. Identifikator først og fremst for indikator"gjenkjennelse" hos statistiker. Kan eksempelvis være parameteren som benyttes i valgtVar for å hente ut tilrettelegging av denne indikatoren.

MaalNivaaGronn

Grense for grønt målnivå. Desimaltall i intervallet 0-1

MaalNivaaGul

Grense for gult målnivå. Desimaltall i intervallet 0-1

MaalRetn

Angir om ønsket målnivået for variabelen er høyt (1) eller lavt (09. Heltall: 0,1

BeskrivelseKort

Kort, men noe detaljert beskrivelse av indikatoren. Tekst.

BeskrivelseLang

Utfyllende beskrivelse av indikatoren. Tekst.

Source

Organisasjonsnumre og struktur kan finnes ved hjelp av Brønnøyaundregisteret: https://w2.brreg.no/enhet/sok/